pyproject.toml
[project]
name = "piano-site"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
]

[tool.poetry]
packages = [{include = "piano_site", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
